==> 1-Extraction.ipynb <==
from pymongo import MongoClient
import pandas as pd
import time
import re

client = MongoClient("localhost:27017")
client.list_database_names()
db = client ["ironhack"]
db.list_collection_names()

data_base = db.get_collection("companies")
data_base

# Selecting companies :
In this notebook, we're going to select companies with criterias relevant to our search, and turn the result we get into a clean dataFrame.

## Starting from the "companies" database in Mongo, and importing only the companies :
 1-that do anything related to gaming design..  
 2- that have total money raised greater than 10M$.
 
 
 
 ( instead of 1M$, to have a smaller sample size)

data_base.distinct("total_money_raised")

condition1 = {"total_money_raised" : {"$regex": "\d{2,}\.?\d?M|\d{1,}\.?\d?B"}}
condition2 = {"tag_list": {"$regex": ".*design.|.tech.|.software.|.fashion.|.game."}}
condition3 = {"category_code": "web"}
condition4= {"offices":{"$ne":[]}}
projection = {"_id": 0, "name":1, "offices.city":1, "offices.latitude":1, "offices.longitude":1,"offices.country_code":1 }
query = {"$and": [condition1, condition4, {"$or": [condition2, condition3]}]}
result = list(data_base.find(query, projection))
              
              
                            

len(result)

df = pd.DataFrame(result)
df

df.offices[0]

df = df.explode("offices")
df

df = df.reset_index()
df

cities = []
latitude = []
longitude = []
country = []

for index, row in df.iterrows():
    try:
        cities.append(row["offices"]["city"])
        latitude.append(row["offices"]["latitude"])
        longitude.append(row["offices"]["longitude"])
        country.append(row["offices"]["country_code"])                
    except IndexError :
        cities.append(None)
        latitude.append(None)
        longitude.append(None)
        country_code.append(None)

df["city"] = cities
df["latitude"]= latitude
df["longitude"]= longitude
df["country_code"]= country


#drop offices
new_df = df.drop(["offices"], axis = 1)
new_df


new_df.duplicated().sum()

new_df.dropna(axis=0, inplace=True, how="any")


new_df

new_df.shape

new_df["country_code"].unique()

city_count = new_df["city"].value_counts()
#print(city_count.to_string())

Deleting rows where we have only 1 location. Because we want a location where there are other companies around.

new_df = new_df[new_df.groupby("city").city.transform("count")>1]
new_df

city_count = new_df["city"].value_counts()
#print(city_count.to_string())

new_df.reset_index()

result1 = new_df.reset_index().rename(columns={'index' : 'col1_index'})
result1

#result1 = result1.drop(["col1_index","level_0"],axis=1)
result1 = result1.drop("level_0",axis=1)

result1


Now we have a data frame with all the potential locations around the globe, that meet some of our requiremenst,  
we're taking this data frame into another Jupyter notebook, for further investigations using APIs calls

result1.to_csv("./dataset.csv", index = False)

df = pd.read_csv("./dataset/dataset.csv")

df.sample(6)
