==> 2-API_queries.ipynb <==
from pymongo import MongoClient
import pandas as pd
import time
import re
import folium
from folium import Choropleth, Circle, Marker, Icon, Map
from folium.plugins import HeatMap, MarkerCluster
from dotenv import load_dotenv
import os
import requests
import warnings
warnings.filterwarnings("ignore")

df = pd.read_csv("./dataset/dataset.csv")

# Filtering the number of companies, according to criterias we have


The filter process was split into 2 steps,  

1- First filter we used a mongo query to select form the database design companies that had raised more than 10M$, analysed in the first Jupyet Notebook.  

2- For the second filter we are going to be using foursquare API to narrow down our dataset, and get a final dataframe with only companies that meet our conditions:  
 -- we will be deleting any company(coordinates) that has no amenities around.  
 -- for that matter we will be creating additional column for each amenity with their number(how many are there)
 
    


load_dotenv()

token_fsq = os.getenv("token")

df.country_code.unique()


#this is a test
usa_df = df[df["country_code"] == "USA"]
#test api with country_code FRA, bc it doesn't have too many values
fra_df = df[df["country_code"] == "FRA"]


# Starbucks
1- look for Starbucks cafes nearby, because everyone needs coffee.  
2- deleting location with no starbucks around.

     (setting the radius to 500 and limiting search to 50.)

def getNearbyStarbucks(lat, long, radius=500):

    # create the API request URL
    url = f"https://api.foursquare.com/v3/places/search?ll={lat}%2C{long}&radius={radius}&chains=ab4c54c0-d68a-012e-5619-003048cad9da&exclude_all_chains=false&sort=DISTANCE&limit=50"        
    # make the GET request
    headers = {
    "accept": "application/json",
    "Authorization": token_fsq
    }
    
    results = requests.get(url,headers=headers).json()["results"]
    
    return len(results)
     

for index, row in df.iterrows():
    df.at[index, "Starbucks"] = getNearbyStarbucks(row["latitude"],row["longitude"])
   

star_count = df["Starbucks"].value_counts()
#print(star_count.to_string())

     


df['Starbucks'].dtypes
df['Starbucks'] = df['Starbucks'].astype(int)

#pd.set_option('display.max_rows', 500)
#df


#deleting the locations with zero Starbucks nearby
df = df[df.Starbucks != 0]

# Nightlife

1- look for Bars nearby, because everyone in the company is between 25 and 40, and need some place to go party.  
2- deleting location with no bars around.  


 (setting the radius to 500 and limiting search to 50.)

def getNearbyNightlife(lat, long, radius=500):

    # create the API request URL
    url = f"https://api.foursquare.com/v3/places/search?ll={lat}%2C{long}&radius={radius}&categories=13003&exclude_all_chains=false&sort=DISTANCE&limit=50"

    # make the GET request
    headers = {
    "accept": "application/json",
    "Authorization": token_fsq
    }
    
    results = requests.get(url,headers=headers).json()["results"]
    
    return len(results)


for index, row in df.iterrows():
    df.at[index, "Nightlife"] = getNearbyNightlife(row["latitude"],row["longitude"])

df['Nightlife'] = df['Nightlife'].astype(int)

nl_count = df["Nightlife"].value_counts()
#print(nl_count.to_string())

df = df[df.Nightlife != 0]

city_count = df["country_code"].value_counts()
#print(city_count.to_string())

# Transportation
Third filter is Transportation, specifically train station,metro station and taxi.(sone of them offer shuttle to the airport)
--assuming that airports are generally located outside cities, by choosing an office near an airport,we are going to be discarding other criterias that can't be found near airports, such as schools,pet grooming services...
 


def getNearbyTransport(lat, long, radius=500):

    # create the API request URL
    url = f"https://api.foursquare.com/v3/places/search?ll={lat}%2C{long}&radius={radius}&categories=19047%2C19049%2C19046&exclude_all_chains=false&sort=DISTANCE&limit=50"

    # make the GET request
    headers = {
    "accept": "application/json",
    "Authorization": token_fsq
    }
    
    results = requests.get(url,headers=headers).json()["results"]
    
    return len(results)

for index, row in df.iterrows():
    df.at[index, "transportation"] = getNearbyTransport(row["latitude"],row["longitude"])

df['transportation'] = df['transportation'].astype(int)

df = df[df.transportation != 0]

transp_count = df["transportation"].value_counts()
#print(transp_count.to_string())
#df

# Preschool
30% of the company staff have at least 1 child.  
Then all the locations with 0 preschool around are going to be deleted.



def getNearbypreschool(lat, long, radius=1000):

    # create the API request URL
    url = f"https://api.foursquare.com/v3/places/search?ll={lat}%2C{long}&radius={radius}&categories=12056&exclude_all_chains=false&sort=DISTANCE&limit=50"

    # make the GET request
    headers = {
    "accept": "application/json",
    "Authorization": token_fsq
    }
    
    results = requests.get(url,headers=headers).json()["results"]
    
    return len(results)
         


for index, row in df.iterrows():
    df.at[index, "preschool"] = getNearbypreschool(row["latitude"],row["longitude"])

df['preschool'] = df['preschool'].astype(int)

ps_count = df["preschool"].value_counts()
#print(ps_count.to_string())

df = df[df.preschool != 0]

city_count = df["country_code"].value_counts()
#print(city_count.to_string())

# Vegan restaurants

def getNearbyVeganRest(lat, long, radius=500):

    # create the API request URL
    url = f"https://api.foursquare.com/v3/places/search?ll={lat}%2C{long}&radius={radius}&categories=13377&exclude_all_chains=false&sort=DISTANCE&limit=50"

    # make the GET request
    headers = {
    "accept": "application/json",
    "Authorization": token_fsq
    }
    
    results = requests.get(url,headers=headers).json()["results"]
    
    return len(results)
      

for index, row in df.iterrows():
    df.at[index, "veganRest"] = getNearbyVeganRest(row["latitude"],row["longitude"])

df['veganRest'] = df['veganRest'].astype(int)

vegan_count = df["veganRest"].value_counts()
#print(vegan_count.to_string())
#df

city_count = df["country_code"].value_counts()
print(city_count.to_string())

df = df[df.veganRest != 0]

# Pet grooming service
The office dog—"Dobby" needs a hairdresser every month. Ensure there's one not too far away.  
the radius is set to 1000, Dobby only needs to go there once a month.


def getNearbyPetService(lat, long, radius=1000):

    # create the API request URL
    url = f"https://api.foursquare.com/v3/places/search?ll={lat}%2C{long}&radius={radius}&categories=11134&exclude_all_chains=false&sort=DISTANCE"

    # make the GET request
    headers = {
    "accept": "application/json",
    "Authorization": token_fsq
    }
    
    results = requests.get(url,headers=headers).json()["results"]
    
    return len(results)
      

for index, row in df.iterrows():
    df.at[index, "petService"] = getNearbyPetService(row["latitude"],row["longitude"])

df['petService'] = df['petService'].astype(int)

Ps_count = df["petService"].value_counts()
#print(Ps_count.to_string())

df = df[df.petService != 0]

city_count = df["country_code"].value_counts()
#print(city_count.to_string())

city_count = df["city"].value_counts()
#print(city_count.to_string())

test1 = df[~df.country_code.str.contains("USA")]
test1


df.to_csv("./dataset/world_dataset.csv", index = False)


new_df = pd.read_csv("./dataset/world_dataset.csv")

# Dataset with all needed amenities

new_df.sample()

Now that we have a dataset with all the criterias we need, we are going to check the total amenities in each location, to have a look at the density : we want our office to be in a dense location.  
1- we are getting the total amenities for each location  
2- we are checking the ration...


new_df["Total"] = (new_df["Starbucks"])+(new_df["Nightlife"])+(new_df["transportation"])+(new_df["preschool"])+(new_df["veganRest"])+(new_df["petService"])


new_df["density"] = (new_df["Starbucks"]*1)+(new_df["Nightlife"]*0.8)+(new_df["transportation"]*1)+(new_df["preschool"]*0.7)+(new_df["veganRest"]*0.6)+(new_df["petService"]*0.6)

new_df = new_df.sort_values(by='density', ascending=False)

new_df.sample(4)

# World Heat map
Now we are plotting all the locations we have in our dataset into a heatmap a see the most dense location, in term of companies


companies_group = folium.FeatureGroup(name=f"Companies: {new_df.shape[0]}")

wolrd_map  = Map(location = ["41.382707", "2.169242"], zoom_start = 2)
HeatMap(data=new_df[["latitude", "longitude"]], 
        radius=25,
       gradient = {0.4: 'yellow', 0.65: 'orange', 1: 'white'}).add_to(companies_group)

companies_group.add_to(wolrd_map)

wolrd_map
wolrd_map.save('wolrd_map.html')

# It looks like we're going to New York !
using the ratio/density feature, we are going to sort the data set in an descending order, because we are looking for the place we the most number of amenities.


ny_df = new_df[new_df["city"] == "New York"]



ny_df.to_csv("./dataset/new_york_data.csv", index = False)

nyork_df = pd.read_csv("./dataset/new_york_data.csv")


#nyork_df


NY_map = Map(location = ["40.751892", "-73.987592"], zoom_start = 10)
folium.Circle([40.751892, -73.987592],radius=1000).add_to(NY_map)


NY_map

#top 10 locations in NY based on density

TOP_NY_map = Map(location = ["40.751892", "-73.987592"], zoom_start = 15)

icon = Icon(color = "blue",
            opacity = 0.1,
            prefix = "fa",
            icon = "briefcase",
            icon_color = "black")

for index, row in top_ny.iterrows():
    
    folium.Circle([row["latitude"],row["longitude"]],radius=1000).add_to(TOP_NY_map)
    folium.Marker(location =[row["latitude"],row["longitude"]],tooltip = row["name"],icon = icon).add_to(TOP_NY_map)
    

TOP_NY_map
