==> 3-Visualization_Maps.ipynb <==
from pymongo import MongoClient
import pandas as pd
import time
import re
import folium
from folium import Choropleth, Circle, Marker, Icon, Map
from folium.plugins import HeatMap, MarkerCluster
from dotenv import load_dotenv
import os
import requests
import warnings
warnings.filterwarnings("ignore")
from pymongo import GEOSPHERE
import json
#from cartoframes.viz import Map, Layer, popup_element
import geopandas as gpd
import geopy.distance


nyork_df = pd.read_csv("./dataset/new_york_data.csv")


load_dotenv()

token_fsq = os.getenv("token")

Now we have a data set with locations in NY, we are going to take the top 3 based on the density, and show them on a map with ametnities around.  
    we're also choosing the top 5 locations that are next to each other, because we want our company to be around other companies that had raised more thn 10M$.  
    
( so far we have been trying to choose a location of a company who had raised 10M\$, so we need to check the surroundings as well)



map2 = Map(location = ["40.751892", "-73.987592"], zoom_start = 15)
#Optimum Energy
folium.Circle([40.751892, -73.987592],radius=1000).add_to(map2)
#Yipit
#folium.Circle([40.744618, -73.987764],radius=1000).add_to(map2)
#Heavy
#folium.Circle([40.753385, -73.989620],radius=1000).add_to(map2)
#gaming major
#folium.Circle([40.752672, -73.975930],radius=1000).add_to(map2)
#tripology
#folium.Circle([40.746883, -73.983969],radius=1000).add_to(map2)


#map2

#add Tripology location

icon = Icon(color = "blue",
            opacity = 0.1,
            prefix = "fa",
            icon = "building",
            icon_color = "black"
)
tripology = Marker(
    location = [40.746883, -73.983969],
    tooltip = "Tripology",
    icon=icon,
    radius = 1000
)
tripology.add_to(map2)

#add Yipit location

icon = Icon(color = "blue",
            opacity = 0.1,
            prefix = "fa",
            icon = "building",
            icon_color = "black"
)
Yipit = Marker(
    location = [40.750597, -73.985273],
    tooltip = "Yipit",
    icon=icon,
    radius = 1000
)
Yipit.add_to(map2)

#add "heavy" location

icon = Icon(color = "blue",
            opacity = 0.1,
            prefix = "fa",
            icon = "building",
            icon_color = "black"
)
Heavy = Marker(
    location = [40.753385, -73.989620],
    tooltip = "Heavy",
    icon=icon,
    radius = 1000
)
Heavy.add_to(map2)

icon = Icon(color = "blue",
            opacity = 0.1,
            prefix = "fa",
            icon = "building",
            icon_color = "black"
)

major_league = Marker(
    location = [40.752672, -73.975930],
    tooltip = "major league",
    icon=icon,
)
major_league.add_to(map2)

#add Optimum_Energy location

icon = Icon(color = "blue",
            opacity = 0.1,
            prefix = "fa",
            icon = "building",
            icon_color = "black",
)
Optimum_Energy = Marker(
    location = [40.751892, -73.987592],
    tooltip = "Optimum Energy",
    icon=icon
)
Optimum_Energy.add_to(map2)

def getNearbyStarbucks(lat, long, radius=1000):

    # create the API request URL
    url = f"https://api.foursquare.com/v3/places/search?ll={lat}%2C{long}&radius={radius}&chains=ab4c54c0-d68a-012e-5619-003048cad9da&exclude_all_chains=false&sort=DISTANCE&limit=10"        
    # make the GET request
    headers = {
    "accept": "application/json",
    "Authorization": token_fsq
    }
    
    response = requests.get(url,headers=headers).json()    
    return response
     

response_star = getNearbyStarbucks("40.751892","-73.987592")

starbucks = []
for i in response_star["results"]:
    
    name = i["name"]
    latitude = i["geocodes"]["main"]["latitude"]
    longitude = i["geocodes"]["main"]["longitude"]
    distance = i["distance"]
    
    starbucks.append({"name":name, "latitude":latitude, "longitude":longitude, "distance_Optimum Energy": distance})
    
    starbucks

df_cafe = pd.DataFrame(starbucks)


for index, row in df_cafe.iterrows():
    
    #1. MARKER without icon
    starbucks = {"location": [row["latitude"], row["longitude"]], "tooltip": row["name"]}
    
    #2. Icon
          
    icon = Icon (
            color="lightred",
            opacity = 0.6,
            prefix = "fa",
            icon="coffee",
            icon_color = "black"
        )
     #3. Marker
    new_marker = Marker(**starbucks, icon = icon, radius = 1000)
    
    #4. Add the Marker
    new_marker.add_to(map2)

def getNearbyLocation(lat, long, categorie, radius=1000):

    # create the API request URL
    url = f"https://api.foursquare.com/v3/places/search?ll={lat}%2C{long}&radius={radius}&categories={categorie}&exclude_all_chains=false&sort=DISTANCE&limit=10"

    # make the GET request
    headers = {
    "accept": "application/json",
    "Authorization": token_fsq
    }    
    response = requests.get(url,headers=headers).json()    
    
    return response

night_life_response = getNearbyLocation("40.751892","-73.987592","13003")

night_life = []
for i in night_life_response["results"]:
    
    name = i["name"]
    latitude = i["geocodes"]["main"]["latitude"]
    longitude = i["geocodes"]["main"]["longitude"]
    distance = i["distance"]
    
    night_life.append({"name":name, "latitude":latitude, "longitude":longitude, "distance_Optimum Energy": distance})
    
    night_life

df_night_life = pd.DataFrame(night_life)


for index, row in df_night_life.iterrows():
    
    #1. MARKER without icon
    nightlife = {"location": [row["latitude"], row["longitude"]], "tooltip": row["name"]}
    
    #2. Icon
          
    icon = Icon (
            color="purple",
            opacity = 0.6,
            prefix = "fa",
            icon="glass",
            icon_color = "black"
        )
     #3. Marker
    new_marker1 = Marker(**nightlife, icon = icon, radius = 1000)
    
    #4. Add the Marker
    new_marker1.add_to(map2)

#transportation
transportation_response = getNearbyLocation("40.751892","-73.987592","19047,19049,19046")#19047 rail
#19049 taxi
#19046 metro


coords_1 = (40.751892,-73.987592)
coords_2 = (40.752206,-73.989137)
print (geopy.distance.geodesic(coords_1, coords_2).km)

transportation = []
for i in transportation_response["results"]:
    
    name = i["name"]
    latitude = i["geocodes"]["main"]["latitude"]
    longitude = i["geocodes"]["main"]["longitude"]
    distance = i["distance"]
    
    transportation.append({"name":name, "latitude":latitude, "longitude":longitude, "distance_Optimum Energy": distance})
    
    transportation



df_transportation = pd.DataFrame(transportation)


for index, row in df_transportation.iterrows():
    
    #1. MARKER without icon
    transportation = {"location": [row["latitude"], row["longitude"]], "tooltip": row["name"]}
    
    #2. Icon
          
    icon = Icon (
            color="yellow",
            opacity = 0.6,
            prefix = "fa",
            icon="car",
            icon_color = "white"
        )
     #3. Marker
    new_marker2 = Marker(**transportation, icon = icon, radius = 1000)
    
    #4. Add the Marker
    new_marker2.add_to(map2)

preschool_response = getNearbyLocation("40.751892","-73.987592","12056")

preschool = []
for i in preschool_response["results"]:
    
    name = i["name"]
    latitude = i["geocodes"]["main"]["latitude"]
    longitude = i["geocodes"]["main"]["longitude"]
    distance = i["distance"]
    
    preschool.append({"name":name, "latitude":latitude, "longitude":longitude, "distance_Optimum Energy": distance})
    
    preschool

df_preschool = pd.DataFrame(preschool)

for index, row in df_preschool.iterrows():
    
    #1. MARKER without icon
    preschool = {"location": [row["latitude"], row["longitude"]], "tooltip": row["name"]}
    
    #2. Icon
          
    icon = Icon (
            color="pink",
            opacity = 0.6,
            prefix = "fa",
            icon="child",
            icon_color = "black"
        )
     #3. Marker
    new_marker3 = Marker(**preschool, icon = icon, radius = 1000)
    
    #4. Add the Marker
    new_marker3.add_to(map2)

vegan_response = getNearbyLocation("40.751892","-73.987592","13377")


vegan_rest = []
for i in vegan_response["results"]:
    
    name = i["name"]
    latitude = i["geocodes"]["main"]["latitude"]
    longitude = i["geocodes"]["main"]["longitude"]
    distance = i["distance"]
    
    vegan_rest.append({"name":name, "latitude":latitude, "longitude":longitude, "distance_Optimum Energy": distance})
    
    vegan_rest

df_vegan_rest = pd.DataFrame(vegan_rest)


for index, row in df_vegan_rest.iterrows():
    
    #1. MARKER without icon
    vegan_rest = {"location": [row["latitude"], row["longitude"]], "tooltip": row["name"]}
    
    #2. Icon
          
    icon = Icon (
            color="lightgreen",
            opacity = 0.6,
            prefix = "fa",
            icon="leaf",
            icon_color = "black"
        )
     #3. Marker
    new_marker4 = Marker(**vegan_rest, icon = icon, radius = 1000)
    
    #4. Add the Marker
    new_marker4.add_to(map2)

pet_response = getNearbyLocation("40.751892","-73.987592","11134")


pet_service = []
for i in pet_response["results"]:
    
    name = i["name"]
    latitude = i["geocodes"]["main"]["latitude"]
    longitude = i["geocodes"]["main"]["longitude"]
    distance = i["distance"]
    
    pet_service.append({"name":name, "latitude":latitude, "longitude":longitude, "distance_Optimum Energy": distance})
    
    pet_service

df_pet_service = pd.DataFrame(pet_service)


for index, row in df_pet_service.iterrows():
    
    #1. MARKER without icon
    pet_service = {"location": [row["latitude"], row["longitude"]], "tooltip": row["name"]}
    
    #2. Icon
          
    icon = Icon (
            color="lightblue",
            opacity = 0.6,
            prefix = "fa",
            icon="paw",
            icon_color = "white"
        )
     #3. Marker
    new_marker5 = Marker(**pet_service, icon = icon, radius = 1000)
    
    #4. Add the Marker
    new_marker5.add_to(map2)

map2

map2.save('map2.html')
